Move function validTransitionAndEnd to workflow_validator.go
